refactor: unify strong typed fragment id (#23649)

+++ src/meta/service/src/serving_service.rs
@@ -59,7 +59,7 @@ impl ServingService for ServingServiceImpl {
             .fragment_job_mapping()
             .await?
             .into_iter()
-            .map(|(fragment_id, job_id)| (fragment_id as u32, job_id as u32))
+            .map(|(fragment_id, job_id)| (fragment_id, job_id as u32))
             .collect();
         Ok(Response::new(GetServingVnodeMappingsResponse {
             fragment_to_table,
